lenscat-backend/pyproject.toml
[tool.poetry]
name = "lenscat-backend"
version = "0.1.0"
description = "Minimal FastAPI backend for Lenscat gallery"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
aioboto3 = "^12.3.0"
pyvips = "^2.2.1"
pillow = "^10.1.0"
orjson = "^3.9.10"
blake3 = "^0.4.1"
aiofiles = "^23.2.1"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
black = "^23.11.0"
mypy = "^1.7.1"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "T20"]
ignore = ["S101", "S608"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
